rename: User 패키지 이동 및 파일 분리 #33
