+++ src/main/java/cmc/delta/domain/user/adapter/in/dto/response/UserActionResultData.java
@@ -1,0 +1,4 @@
+package cmc.delta.domain.user.adapter.in.dto.response;
+
+public record UserActionResultData(boolean success) {
+}
