+++ src/main/java/cmc/delta/domain/user/adapter/in/dto/response/UserMeData.java
@@ -1,4 +1,4 @@
-package cmc.delta.domain.user.adapter.in.web.dto.response;
+package cmc.delta.domain.user.adapter.in.dto.response;
 
 public record UserMeData(
 	Long userId,
