--- src/main/java/cmc/delta/domain/user/adapter/in/web/dto/response/UserActionResultData.java
@@ -1,4 +1,0 @@
-package cmc.delta.domain.user.adapter.in.web.dto.response;
-
-public record UserActionResultData(boolean success) {
-}
