+++ src/main/java/cmc/delta/domain/user/application/port/in/UserUseCase.java
@@ -1,6 +1,6 @@
 package cmc.delta.domain.user.application.port.in;
 
-import cmc.delta.domain.user.adapter.in.web.dto.response.UserMeData;
+import cmc.delta.domain.user.adapter.in.dto.response.UserMeData;
 
 public interface UserUseCase {
 	UserMeData getMyProfile(long userId);
